[EDELIVERY-2368] null value check for expirationLimit
Allowing to disable expiration limit

+++ Domibus-default-fs-plugin/src/main/java/eu/domibus/plugin/fs/worker/FSAbstractPurgeService.java
@@ -66,7 +66,7 @@ public abstract class FSAbstractPurgeService {
 
     private void checkAndPurge(FileObject file, Integer expirationLimit) {
         try {
-            if (isFileOlder(file, expirationLimit)) {
+            if (expirationLimit != null && isFileOlder(file, expirationLimit)) {
                 LOG.debug("File [{}] is too old. Deleting", file.getName());
                 fsFilesManager.deleteFile(file);
             } else {
